export into empty directory; prime .gitignore from .cvsignore
There are no invariants that ensure that no files will exist in the
Git module that would break cvs export, and cvs export fails if there
are any conflicting files.  Therefore, cvs export into an empty
directory and copy the files over to the git directory afterward.

If there is no skeleton and there is a root-level .cvsignore, create
a root-level .gitignore file that is a copy of the .cvsignore file as
a starting point.