Use -kk option to cvs checkout and cvs export
The -kk option on export means that it will ignore the contents of
keywords, and therefore a change in keywords will not cause a commit
to happen; previously the keywords were causing a commit of a new
revision with zero changes.  The -kk option on export means that sed
is only needed for disabling the Log keyword so that it will not
generate additional merge conflicts.