discard local changes when switching branches
There are three ways that we could have outstanding changes in a
the Git local working directory when we call "git checkout" to
change the current branch.

1. User has modified the working directory externally.
2. This program crashed/aborted with a modified working directory.
3. Line ending normalization has occurred.

In each of those cases, we do want the checkout to succeed, and
explicitly to throw away any local modifications, rather than
make additional commits of unknown content, or abort due to
uncommitted changes in the working directory.  Therefore, it is
appropriate to use the -f (force) option to "git checkout".