cvs checkout into the correct directory
Modules were previously checked out from CVS into a directory
named for the Git repository, but then all other code expected
all the files to exist in a directory named by the CVS repository.
It has been fixed to consistently use the CVS module name for the
CVS checkout directory.  The only impact of this bug was when the
basename of the CVS module and the Git repository were different.