refactor git->cvs export: faster, remove wrong "initial commit" messages
The Git->CVS export is refactored to support more unit testing, and
unit tests are added.

Previously, if there had been no changes in Git, but the CVS server
had renormalized files, the normalized files would be re-committed
(with the same content; no changes) with an incorrect log message
saying that it was the initial commit.  Now, if there have been no
changes in Git, the export operation terminates, and there is no
wrong commit message created.

This change has two operational impacts.

The export process does not have to do a CVS checkout or update
operation when Git hasn't changed, so it is faster when there are
no Git changes to push.

Re-running the export process will no longer overwrite changes made
in CVS, unless an additional change has been committed in Git on
the relevant branch.