pre-commit and post-commit hooks
In order to enable arbitrary modifications during the import/export
process (for example, normalizing line endings or enforcing standards)
pre-commit hooks for Git and CVS are added.  Similarly, in order to
enable arbitrary notifications after committing, post-commit hooks for
Git and CVS are added.

These hooks can be global or per repository, and can be for all
branches or branch-specific.  The hooks are enabled in repository
configuration.