use fewer tracking branches
When line ending normalization is in place, it is possible for Git
to refuse to do a fast-forward merge only because there are line
ending changes in the working directory of an otherwise clean checkout.
This can break the attempt to fast-forward an export- branch.  Instead,
pushing the exported branch to export-branchname at the origin will
have the same semantics, and the bare repository cannot have the same
unclean-working-directory problem.