add Git hooks that run on export to CVS
Changed default Git hooks to run when exporting to CVS.

Added specific types of Git hooks that run only on export
to CVS or only on import from CVS for the cases where that
matters.

This allows normalization, assertions, or reporting based on
the contents of the Git repository.  It is not meant to
encourage committing changes to the Git repository.