emailing errors implemented
Errors are collected per repository, and after repository processing
is done, if there were errors and email recipients are listed for
a repository, an email is sent to those recipients with specific
errors as well as the entire logs.

The entire error log is no longer logged to the console on errors.
It is still kept in log files, and now sent in email, so the entire
log on the console was too much noise.