initial error reporting for missing CVS target location/branch
The cvs checkout command will do nothing and report success when
asked to check out a directory that has no files on the requested
branch.  It does this even when the -P option is not provided.
For now, recognize this condition and raise a meaningful error
at the time it occurs.