disabling CVS Log keyword changed to a hook
The CVS Log keyword was previously always disabled via a sed command
to avoid merge failures.  However, for large repositories, that sed
command can take significant time to run.  Therefore, it has been
replaced by a hook that can be used if desired but not run when the
Log keyword has not been used.