Send error emails only in case of error
Previously, error email was sent if there was any content in the
standard error log.  However, this was true any time any commands
were run because all command invocations are logged in the error
log.  Now, error email is sent only if attachments with error data
have been attached to the mail.