add bigitrd daemon to sync sets of repositories
The bigitrd daemon synchronizes repositories according to a
daemon configuration file that specifies sets of repository
configuration files to read with various application configuration
files, and then synchronize all the repositories.  It can be
configured with separate wait times for synchronizing only if
changes are found in Git repositories and for synchronizing
regardless of whether changes have been made in Git repositories
(in case changes need to be imported from CVS).

This adds dependencies on daemon (python-daemon) and lockfile
(python-lockfile).