show daemon progress in proctitle and non-detached output
The daemon now shows its current state (which module it is currently
syncing or polling, or the length of time it is sleeping) in proctitle,
and if not detached, in output.  This adds a requirement for
python-setproctitle.

Lock handling now works, including breaking the lock if no daemon
process is found, and reporting the process holding the lock if a
daemon process is found.  Restarting on SIGHUP now works.  Handles
Control-C (SIGINT) the same way as SIGTERM, stopping after current
job finishes.

Added story testing for the daemon, including bidirectional sync,
SIGHUP, and SIGTERM handling.