forward to remote branch unconditionally (issues/6)
Previously, JGit ignoring eol specification in .gitattributes could
cause failures during operations in subsequent cycles, normally while
attempting to synchronize the fix for incorrect line termination.
The correct solution is to reset the HEAD to the remote branch
which has the effect of a "forced fast-forward merge" (if such a
thing existed).