remove potentially inconsistent CVS checkouts (issues/8)
There are potential causes for CVS commands to return errors that
the client cannot check for.  For instance, creating a file with a
the same name as a directory that has ever been created, on any
branch in the repository, will cause the commit to fail.  These
failing CVS commands can leave the CVS checkout in an inconsistent
state.

When a CVS command terminates with an unexpected error, remove the
CVS checkout so that future actions will be trustworthy.

Fixes #8