No longer delete auto-ignored files in some cases
Previously, deletions caused by .gitignore and .cvsignore were mirrored
to the other system on sync; if a file in CVS was not in Git due to a
line in .gitignore, when Git was exported to CVS, the files was deleted.

Now, for a subset of files and specifications documented in README.md,
a file matched in .gitignore will not be deleted from CVS on export
from Git to CVS, and a file matched in .cvsignore will not be deleted
from Git on import from CVS into Git.

Fixes #16