Clarify contributor agreement
It was always the intention that the contributor agreement apply
to any contribution.  However, the word "change" was confusing to
some readers, and seemed to imply a distinction between new features
and modifications to existing features.  To clarify the intent,
the agreement has been modified to explicitly apply to all
contributions, regardless of whether the person providine the
contribution considers it a "change".