Implement .bigitrsync in CVS to allow partial sync
If a .bigitrsync file is present in the root of a CVS module to
which a Git repository is being exported, then it contains regular
expressions, one per line, defining file names that will be synced
from Git to CVS.  Files present in CVS but not in Git will be deleted
from CVS only if they do not match one of the sync patterns in the
.bigitrsync file.

If a .bigitrsync file is present in the root of either a CVS or a
Git repository, it will not be synced.

A .bigitrsync file in a Git repository does not currently implement
the equivalent limitation for import from CVS into Git, but the file
name is reserved to perform such a limitation in the future if there
is ever a use case for doing so.