Allow custom git log options for CVS messages
When using git log to export Git commits to CVS, allow configuration
to specify extra command-line options to the git log command to
put branch-specific additional information in the CVS commits.
The configuration is specific to a Git branch of a repository,
and is not looked up in global configuration:

    [repository]
    gitlog.<branchname> = --list --of --arguments

There is NO corresponding general "gitlog" configuration option
to specify command arguments for all branches.